Update add-to-function-php
Hide ACF (Advanced Custom Fields) in WordPress Admin in Production Mode

This repository contains a code snippet for WordPress, specifically tailored for those using the Bedrock and Sage frameworks by Roots.io. The snippet removes the ACF menu from the admin sidebar and admin bar, and restricts direct access to ACF pages when in production mode. This is useful for securing custom field management and preventing changes in the production environment.